feat(lua)!: methods/commands to add and remove dependencies form a RTD
Add the following methods to the RuntimeDir class:

- `add_dependencies(rtds)`

- `remove_dependencies(rtds)`

The `deps` attribute of the `RuntimeDir` class is no longer a object
with the `rtds` and `names` attributes. It is an array with all runtime
directories that are dependencies of the current one. To access its
names, use the `config` attribute

Add user commands to add and remove dependencies to/from a runtime
directory:

- 'ProjectRtdAddRtdDependencies'

- 'ProjectRtdRemoveRtdDependencies'

BREAKING CHANGE: Change the type of the `deps` attribute of the
`RuntimeDir``